notification/requirements.txt
asgiref==3.9.2
bidict==0.23.1
click==8.3.0
Django==5.2.7
django-environ==0.12.0
djangorestframework==3.16.1
dnspython==2.8.0
eventlet==0.40.3
greenlet==3.2.4
h11==0.16.0
psycopg2-binary==2.9.10
python-engineio==4.12.3
python-socketio==5.14.0
simple-websocket==1.1.0
sqlparse==0.5.3
typing_extensions==4.15.0
uvicorn==0.37.0
wsproto==1.2.0
